Move terminal resize overlay to TerminalInstance; support multi-terminal and manual resize
